Add permissions to group

+++ app/Http/Controllers/Api/ContactGroup/ContactGroupController.php
@@ -35,6 +35,8 @@ class ContactGroupController extends Controller
 
     public function store(RequestInput $requestInput)
     {
+        $this->authorize('createGroup', ContactGroup::class);
+
         $data = $requestInput->string('name')->whenMissing('')->next()
             ->string('description')->whenMissing('')->next()
             ->boolean('closed')->validate('boolean')->whenMissing(false)->next()
@@ -51,6 +53,8 @@ class ContactGroupController extends Controller
 
     public function update(RequestInput $requestInput, ContactGroup $contactGroup)
     {
+        $this->authorize('editGroup', ContactGroup::class);
+
         $data = $requestInput->string('name')->next()
             ->string('description')->next()
             ->boolean('closed')->validate('boolean')->next()
@@ -67,6 +71,8 @@ class ContactGroupController extends Controller
 
     public function destroy(ContactGroup $contactGroup)
     {
+        $this->authorize('deleteGroup', ContactGroup::class);
+
         $contactGroup->delete();
     }
 
@@ -77,16 +83,20 @@ class ContactGroupController extends Controller
 
     public function addContact(ContactGroup $contactGroup, Contact $contact)
     {
+        $this->authorize('addToGroup', ContactGroup::class);
         $contactGroup->contacts()->attach($contact);
     }
 
     public function removeContact(ContactGroup $contactGroup, Contact $contact)
     {
+        $this->authorize('removeFromGroup', ContactGroup::class);
         $contactGroup->contacts()->detach($contact);
     }
 
     public function addContacts(ContactGroup $contactGroup, Request $request)
     {
+        $this->authorize('addToGroup', ContactGroup::class);
+
         $contactIds = json_decode($request->input('ids'));
 
         $contactGroup->contacts()->attach($contactIds);
